split grade interface file
