+++ polymorphia-frontend/hooks/course/useGradeUpdate/types.ts
@@ -1,4 +1,4 @@
-import { GradeRequestDTO } from "@/interfaces/api/grade";
+import { GradeRequestDTO } from "@/interfaces/api/grade/grade";
 
 export interface UseGradeUpdate {
   mutate: (variables: GradeRequestDTO) => void;
