+++ polymorphia-frontend/hooks/course/useGradingTargets/index.tsx
@@ -3,7 +3,7 @@ import { EventSectionService } from "@/app/(logged-in)/course/EventSectionServic
 import { useEventParams } from "@/hooks/general/useEventParams";
 import { UseGradingTargets } from "./types";
 import { EventTypes } from "@/interfaces/general";
-import { TargetTypes } from "@/interfaces/api/grade";
+import { TargetTypes } from "@/interfaces/api/grade/target";
 
 export default function useGradingTargets(
   debouncedSearch: string,
