+++ polymorphia-frontend/hooks/course/useGradingTargets/types.ts
@@ -1,4 +1,4 @@
-import { TargetResponseDTO } from "@/interfaces/api/grade";
+import { TargetResponseDTO } from "@/interfaces/api/grade/target";
 
 export interface UseGradingTargets {
   data: TargetResponseDTO[] | undefined;
