+++ polymorphia-frontend/hooks/course/useRandomPeopleWithPoints/types.ts
@@ -1,4 +1,4 @@
-import { StudentTargetData } from "@/interfaces/api/grade";
+import { StudentTargetData } from "@/interfaces/api/grade/target";
 
 export interface UseRandomPeopleWithPoints {
   data: StudentTargetData[] | undefined;
